Refactor UI code: Modularize DOM element references and separate shop screen logic
- Moved DOM element references to a new file `domElements.ts` for better organization.
- Created a new file `shopScreen.ts` to encapsulate shop-related UI logic.
- Updated `ui.ts` to re-export functions from the new modularized files.
- Refactored perk and save management code to improve readability and maintainability.
