+++ ts/ui/domElements.ts
@@ -1,0 +1,13 @@
+// ts/ui/domElements.ts
+
+export const fpsCounterEl = document.getElementById('fps-counter') as HTMLElement;
+export const scoreEl = document.getElementById('score') as HTMLElement;
+export const coinsEl = document.getElementById('coins') as HTMLElement;
+export const waveEl = document.getElementById('wave') as HTMLElement;
+export const modalContainer = document.getElementById('modal-container') as HTMLElement;
+export const modalContent = document.getElementById('modal-content-main') as HTMLElement;
+export const pauseButton = document.getElementById('pause-button') as HTMLElement;
+export const pauseIcon = document.getElementById('pause-icon') as HTMLElement;
+export const bossUiContainer = document.getElementById('boss-ui-container') as HTMLElement;
+export const bossNameEl = document.getElementById('boss-name') as HTMLElement;
+export const bossHealthBarEl = document.getElementById('boss-health-bar') as HTMLElement;
